Exit codes and error-chain error handling
Following the work done in https://github.com/resin-io-modules/network-manager/issues/109, now error-chain error handling is introduced in WiFi Connect as well. This simplifies the error handling code and organizes it better.

But most importantly when an error occurs the underlying causes are printed in a nicer way:
```
$ ./wifi-connect -i notexisting
Error: Cannot find network device with interface name 'notexisting'
  caused by: D-Bus failure: org.freedesktop.NetworkManager::GetDeviceByIpIface method call failed on /org/freedesktop/NetworkManager
  caused by: "No device found for the requested iface."

$ echo $?
10
```

In the above case, the first error string is generated in WiFi Connect, the second one in our network-manager bindings, and the last third one in NetworkManager itself (sent using the D-Bus API). Additionally the line starting with `Error:` is colored in red in the terminal for better visibility.

In comparison the previous version produced:
```
$ ./wifi-connect -i notexisting
[wifi_connect:ERROR] D-Bus 'org.freedesktop.NetworkManager'::'GetDeviceByIpIface' method call failed on '/org/freedesktop/NetworkManager': No device found for the requested iface.

$ echo $?
0
```

We now support exit codes for each kind of error defined in WiFi Connect as visible from the examples above.

Change-Type: patch
Connects-To: #166